feat: update various components to use renamed constants and add new blog pages
This commit applies a number of changes across the application. It updates various components (such as the Footer, FeaturesSection, and Menu) to use the renamed CONTENT_HOMEPAGE constant instead of the old CONTENT one. It also adds new pages under the 'src/app' directory for the blog, about, and collection sections of the site. Adjustments are made to the ProductCard component to make the description and price fields optional. The 404 - Page Not Found display is updated for better centering. Finally, it introduces a new BlogSection component and a new blog post page.